+++ authority-portal-frontend/src/app/core/api/fake-backend/impl/fake-organizations.ts
@@ -1,7 +1,24 @@
-import {OwnOrganizationDetailsDto} from '@sovity.de/authority-portal-client';
+import {
+  IdResponse,
+  OrganizationDetailsDto,
+  OrganizationOverviewEntryDto,
+  OrganizationOverviewResult,
+} from '@sovity.de/authority-portal-client';
 import {Patcher, patchObj} from 'src/app/core/utils/object-utils';
 
-export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
+export const approveOrganization = (mdsId: string): IdResponse => {
+  updateOrganization(mdsId, () => ({registrationStatus: 'ACTIVE'}));
+
+  return {id: mdsId, changedDate: new Date()};
+};
+
+export const rejectOrganization = (mdsId: string): IdResponse => {
+  updateOrganization(mdsId, () => ({registrationStatus: 'REJECTED'}));
+
+  return {id: mdsId, changedDate: new Date()};
+};
+
+export let TEST_ORGANIZATIONS: OrganizationDetailsDto[] = [
   {
     mdsId: 'MDSL1111AA',
     name: 'Dev Organization 1',
@@ -14,29 +31,25 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     memberInfos: [
       {
         userId: '00000000-0000-0000-0000-00000001',
-        firstName: 'John',
-        lastName: 'Doe',
-        roles: ['AUTHORITY_ADMIN', 'AUTHORITY_USER'],
+        firstName: 'Authority',
+        lastName: 'Admin',
+        roles: [
+          'AUTHORITY_ADMIN',
+          'AUTHORITY_USER',
+          'PARTICIPANT_ADMIN',
+          'PARTICIPANT_USER',
+        ],
       },
       {
         userId: '00000000-0000-0000-0000-00000002',
-        firstName: 'John',
-        lastName: 'Doe',
-        roles: ['PARTICIPANT_ADMIN'],
-      },
-      {
-        userId: '00000000-0000-0000-0000-00000003',
-        firstName: 'Alice',
-        lastName: 'Smith',
-        roles: ['PARTICIPANT_ADMIN'],
-      },
-      {
-        userId: '00000000-0000-0000-0000-00000004',
-        firstName: 'Bob',
-        lastName: 'Johnson',
-        roles: ['PARTICIPANT_USER'],
+        firstName: 'Authority',
+        lastName: 'User',
+        roles: ['AUTHORITY_USER', 'PARTICIPANT_USER'],
       },
     ],
+    memberCount: 2,
+    connectorCount: 1,
+    dataOfferCount: 0,
   },
 
   {
@@ -50,32 +63,22 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     createdAt: new Date('2023-08-06T00:00:00.000Z'),
     memberInfos: [
       {
-        userId: '00000000-0000-0000-0000-00000001',
-        firstName: 'John',
-        lastName: 'Doe',
-        roles: ['AUTHORITY_ADMIN', 'AUTHORITY_USER'],
-      },
-      {
-        userId: '00000000-0000-0000-0000-00000002',
-        firstName: 'John',
-        lastName: 'Doe',
-        roles: ['PARTICIPANT_ADMIN'],
-      },
-      {
         userId: '00000000-0000-0000-0000-00000003',
-        firstName: 'Alice',
-        lastName: 'Smith',
-        roles: ['PARTICIPANT_ADMIN'],
+        firstName: 'Participant',
+        lastName: 'Admin',
+        roles: ['PARTICIPANT_ADMIN', 'PARTICIPANT_CURATOR', 'PARTICIPANT_USER'],
       },
       {
         userId: '00000000-0000-0000-0000-00000004',
-        firstName: 'Bob',
-        lastName: 'Johnson',
+        firstName: 'Participant',
+        lastName: 'User',
         roles: ['PARTICIPANT_USER'],
       },
     ],
+    memberCount: 2,
+    connectorCount: 3,
+    dataOfferCount: 0,
   },
-
   {
     mdsId: 'MDSL3333CC',
     name: 'Test Orga',
@@ -87,30 +90,21 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     createdAt: new Date('2023-08-01T00:00:00.000Z'),
     memberInfos: [
       {
-        userId: '00000000-0000-0000-0000-00000001',
-        firstName: 'John',
-        lastName: 'Doe',
-        roles: ['AUTHORITY_ADMIN', 'AUTHORITY_USER'],
-      },
-      {
-        userId: '00000000-0000-0000-0000-00000002',
+        userId: '00000000-0000-0000-0000-00000012',
         firstName: 'John',
         lastName: 'Doe',
         roles: ['PARTICIPANT_ADMIN'],
       },
       {
-        userId: '00000000-0000-0000-0000-00000003',
-        firstName: 'Alice',
-        lastName: 'Smith',
-        roles: ['PARTICIPANT_ADMIN'],
-      },
-      {
-        userId: '00000000-0000-0000-0000-00000004',
-        firstName: 'Bob',
-        lastName: 'Johnson',
+        userId: '00000000-0000-0000-0000-00000013',
+        firstName: 'Jack',
+        lastName: 'Doe',
         roles: ['PARTICIPANT_USER'],
       },
     ],
+    memberCount: 2,
+    connectorCount: 1,
+    dataOfferCount: 0,
   },
 
   {
@@ -122,7 +116,23 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     url: 'https://example31.com',
     securityEmail: 'security@example31.com',
     createdAt: new Date('2022-10-01T00:00:00.000Z'),
-    memberInfos: [],
+    memberInfos: [
+      {
+        userId: '00000000-0000-0000-0000-00000022',
+        firstName: 'John',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_ADMIN'],
+      },
+      {
+        userId: '00000000-0000-0000-0000-00000023',
+        firstName: 'Jack',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_USER'],
+      },
+    ],
+    memberCount: 2,
+    connectorCount: 1,
+    dataOfferCount: 0,
   },
   {
     mdsId: 'MDSL3332C2',
@@ -133,7 +143,23 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     url: 'https://example32.com',
     securityEmail: 'security@example32.com',
     createdAt: new Date('2022-10-02T00:00:00.000Z'),
-    memberInfos: [],
+    memberInfos: [
+      {
+        userId: '00000000-0000-0000-0000-00000032',
+        firstName: 'John',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_ADMIN'],
+      },
+      {
+        userId: '00000000-0000-0000-0000-00000033',
+        firstName: 'Jack',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_USER'],
+      },
+    ],
+    memberCount: 2,
+    connectorCount: 1,
+    dataOfferCount: 0,
   },
   {
     mdsId: 'MDSL3333C3',
@@ -144,7 +170,23 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     url: 'https://example33.com',
     securityEmail: 'security@example33.com',
     createdAt: new Date('2022-10-03T00:00:00.000Z'),
-    memberInfos: [],
+    memberInfos: [
+      {
+        userId: '00000000-0000-0000-0000-00000042',
+        firstName: 'John',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_ADMIN'],
+      },
+      {
+        userId: '00000000-0000-0000-0000-00000043',
+        firstName: 'Jack',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_USER'],
+      },
+    ],
+    memberCount: 2,
+    connectorCount: 1,
+    dataOfferCount: 0,
   },
   {
     mdsId: 'MDSL3334C4',
@@ -155,13 +197,29 @@ export let TEST_ORGANIZATIONS: OwnOrganizationDetailsDto[] = [
     url: 'https://example34.com',
     securityEmail: 'security@example34.com',
     createdAt: new Date('2022-10-04T00:00:00.000Z'),
-    memberInfos: [],
+    memberInfos: [
+      {
+        userId: '00000000-0000-0000-0000-00000052',
+        firstName: 'John',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_ADMIN'],
+      },
+      {
+        userId: '00000000-0000-0000-0000-00000053',
+        firstName: 'Jack',
+        lastName: 'Doe',
+        roles: ['PARTICIPANT_USER'],
+      },
+    ],
+    memberCount: 2,
+    connectorCount: 1,
+    dataOfferCount: 0,
   },
 ];
 
 export const updateOrganization = (
   mdsId: string,
-  patcher: Patcher<OwnOrganizationDetailsDto>,
+  patcher: Patcher<OrganizationDetailsDto>,
 ) => {
   TEST_ORGANIZATIONS = TEST_ORGANIZATIONS.map((organization) => {
     return organization.mdsId === mdsId
@@ -169,3 +227,31 @@ export const updateOrganization = (
       : organization;
   });
 };
+
+export const getOrganizationDetails = (
+  mdsId: string,
+): OrganizationDetailsDto => {
+  return TEST_ORGANIZATIONS.find(
+    (organization) => organization.mdsId === mdsId,
+  ) as OrganizationDetailsDto;
+};
+
+export const getOrganizations = (): OrganizationDetailsDto[] => {
+  return TEST_ORGANIZATIONS;
+};
+
+export const getListOfOrganizationsForTable =
+  (): OrganizationOverviewResult => {
+    return {
+      organizations: TEST_ORGANIZATIONS.map(
+        (organization: OrganizationDetailsDto) => {
+          return {
+            mdsId: organization.mdsId,
+            name: organization.name,
+            registrationStatus: organization.registrationStatus,
+            url: organization.url,
+          } satisfies OrganizationOverviewEntryDto;
+        },
+      ),
+    };
+  };