+++ authority-portal-frontend/src/app/pages/participant-own-connector-list-page/state/participant-own-connector-list-page-actions.ts
@@ -3,3 +3,8 @@ const tag = 'ParticipantOwnConnectorListPage';
 export class GetOwnOrganizationConnectors {
   static readonly type = `[${tag}]  Get Own Organization Connectors`;
 }
+
+export class DeleteOwnConnector {
+  static readonly type = `[${tag}]  Delete Own Connector`;
+  constructor(public connectorId: string) {}
+}